STEP 4. Add CRUD for task model

+++ app/Http/Controllers/TaskStatusController.php
@@ -9,37 +9,21 @@ class TaskStatusController extends Controller
 {
     public function __construct()
     {
-        // $this->middleware('auth')->except('index');
         $this->authorizeResource(TaskStatus::class);
     }
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index()
     {
         $taskStatuses = TaskStatus::orderBy('id')->paginate(20);
         return view('taskStatuses.index', compact('taskStatuses'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function create()
     {
         $taskStatuses = new TaskStatus();
         return view('taskStatuses.create', compact('taskStatuses'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(Request $request)
     {
         $data = $this->validate($request, [
@@ -55,24 +39,11 @@ class TaskStatusController extends Controller
         return redirect()->route('task_statuses.index');
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit(TaskStatus $taskStatus)
     {
         return view('taskStatuses.edit', compact('taskStatus'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function update(Request $request, TaskStatus $taskStatus)
     {
         $data = $this->validate($request, [
@@ -86,20 +57,14 @@ class TaskStatusController extends Controller
         return redirect()->route('task_statuses.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function destroy(TaskStatus $taskStatus)
     {
-        // if (!$taskStatus->task()->exists()) {
+        if ($taskStatus->tasks->count() === 0) {
             $taskStatus->delete();
             flash(__('messages.Status deleted successfully!'))->success();
-        // } else {
-        //     flash(__('messages.Action is not possible!'))->warning();
-        // }
+        } else {
+            flash(__('messages.Failed to delete status'))->danger();
+        }
 
         return redirect()->route('task_statuses.index');
     }
